+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * Convert static variables into fields [#623](https://github.com/icsharpcode/CodeConverter/issues/623)
 * Ensure query syntax join conditions are swapped to the necessary C# order [#752](https://github.com/icsharpcode/CodeConverter/issues/752)
 * Convert nested exit statements to if statements [#690](https://github.com/icsharpcode/CodeConverter/issues/690)
+* Convert Exit Try to a do while false loop with a break statement [#779](https://github.com/icsharpcode/CodeConverter/issues/779)
 
 ### C# -> VB